fix: duration int -> bigint

+++ sql/schema/schema.sql
@@ -761,7 +761,7 @@ CREATE TABLE "WorkflowRun" (
     "parentId" UUID,
     "parentStepRunId" UUID,
     "additionalMetadata" JSONB,
-    "duration" INTEGER,
+    "duration" BIGINT,
     "priority" INTEGER,
 
     CONSTRAINT "WorkflowRun_pkey" PRIMARY KEY ("id")
